feat: Manually implementing 'libbpf_set_print' to solve the error problem of platform 'va_list'
