+++ build.rs
@@ -70,7 +70,6 @@ fn generate_bindings(src_dir: path::PathBuf) {
         .allowlist_function("perf_.+")
         .allowlist_function("ring_buffer_.+")
         .allowlist_function("user_ring_buffer_.+")
-        .allowlist_function("vdprintf")
         .allowlist_type("bpf_.+")
         .allowlist_type("btf_.+")
         .allowlist_type("xdp_.+")
@@ -79,6 +78,10 @@ fn generate_bindings(src_dir: path::PathBuf) {
         .allowlist_var("BTF_.+")
         .allowlist_var("XDP_.+")
         .allowlist_var("PERF_.+")
+        .allowlist_type("__va_list_tag")
+        .blocklist_type("vdprintf")
+        .blocklist_type("libbpf_print_fn_t")
+        .blocklist_function("libbpf_set_print")
         .parse_callbacks(Box::new(ignored_macros))
         .header("bindings.h")
         .clang_arg(format!("-I{}", src_dir.join("libbpf/include").display()))
